backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
anthropic==0.21.0
openai==1.3.0
httpx==0.25.2
python-dotenv==1.0.0
